fix some bug

+++ Bookstore/Bookstore/UserControls/Manager_Account_UserControl.xaml.cs
@@ -64,9 +64,16 @@ namespace Bookstore.UserControls
             }
             else
             {
-                db.deleteAccountByID(int.Parse(selectedItem.ID));
-                ManagerAccount_ListView.ItemsSource = managerAccount.Show();
-                MessageBox.Show("Delete success");
+                if (int.Parse(selectedItem.ID) == 1)
+                {
+                    MessageBox.Show("This is super admin you cant delete!!");
+                }
+                else
+                {
+                    db.deleteAccountByID(int.Parse(selectedItem.ID));
+                    ManagerAccount_ListView.ItemsSource = managerAccount.Show();
+                    MessageBox.Show("Delete success");
+                }
             }
         }
 
@@ -93,9 +100,16 @@ namespace Bookstore.UserControls
             }
             else
             {
-                db.deleteAccountByID(int.Parse(selectedItem.ID));
-                ManagerAccount_ListView.ItemsSource = managerAccount.Show();
-                MessageBox.Show("Delete success");
+                if(int.Parse(selectedItem.ID)==1)
+                {
+                    MessageBox.Show("This is super admin you cant delete!!");
+                }
+                else
+                {
+                    db.deleteAccountByID(int.Parse(selectedItem.ID));
+                    ManagerAccount_ListView.ItemsSource = managerAccount.Show();
+                    MessageBox.Show("Delete success");
+                }
             }
         }
 
@@ -114,5 +128,51 @@ namespace Bookstore.UserControls
             ManagerAccount_ListView.ItemsSource = resultAccounts;
         }
 
+
+        GridViewColumnHeader _lastHeaderClicked = null;
+        ListSortDirection _lastDirection = ListSortDirection.Ascending;
+        private void ManagerAccount_ListView_Click(object sender, RoutedEventArgs e)
+        {
+            GridViewColumnHeader headerClicked = e.OriginalSource as GridViewColumnHeader;
+            ListSortDirection direction;
+
+            if (headerClicked != null)
+            {
+                if (headerClicked.Role != GridViewColumnHeaderRole.Padding)
+                {
+                    if (headerClicked != _lastHeaderClicked)
+                    {
+                        direction = ListSortDirection.Ascending;
+                    }
+                    else
+                    {
+                        if (_lastDirection == ListSortDirection.Ascending)
+                        {
+                            direction = ListSortDirection.Descending;
+                        }
+                        else
+                        {
+                            direction = ListSortDirection.Ascending;
+                        }
+                    }
+
+                    string header = headerClicked.Column.Header as string;
+                    Sort(header, direction);
+
+                    _lastHeaderClicked = headerClicked;
+                    _lastDirection = direction;
+                }
+            }
+        }
+
+        private void Sort(string sortBy, ListSortDirection direction)
+        {
+            ICollectionView dataView = CollectionViewSource.GetDefaultView(ManagerAccount_ListView.ItemsSource);
+
+            dataView.SortDescriptions.Clear();
+            SortDescription sd = new SortDescription(sortBy, direction);
+            dataView.SortDescriptions.Add(sd);
+            dataView.Refresh();
+        }
     }
 }
